+++ app/api/show-api/src/main/java/com/example/show/controller/ShowAdminController.java
@@ -1,4 +1,4 @@
-package com.example.controller.show;
+package com.example.show.controller;
 
 public class ShowAdminController {
